requirements.txt
Django==1.7.1
Pillow==2.6.1
argparse==1.2.1
distribute==0.6.24
django-endless-pagination==2.0
django-taggit==0.12.2
django-tinymce==1.5.3
wsgiref==0.1.2
